tag release 7.24.4, a special release for arl that includes support for datum sets (points, axes, planes) and build system fixes

+++ src/libbn/tabdata.c
@@ -1,7 +1,7 @@
 /*                       T A B D A T A . C
  * BRL-CAD
  *
- * Copyright (c) 2004-2014 United States Government as represented by
+ * Copyright (c) 2004-2013 United States Government as represented by
  * the U.S. Army Research Laboratory.
  *
  * This library is free software; you can redistribute it and/or
@@ -18,18 +18,37 @@
  * information.
  */
 
+/** @addtogroup anim */
+/** @{ */
+/** @file libbn/tabdata.c
+ *
+ * @brief
+ *  Routines for processing tables (curves) of data with one independent
+ *  parameter which is common to many sets of dependent data values.
+ *
+ *  Operates on bn_table (independent var) and
+ *  bn_tabdata (dependent variable) structures.
+ *
+ *  One application is for storing spectral curves, see spectrum.c
+ *
+ *  @par Inspired by -
+ *	Roy Hall and his book "Illumination and Color in Computer
+ *@n	Generated Imagery", Springer Verlag, New York, 1989.
+ *@n	ISBN 0-387-96774-5
+ *
+ *  With thanks to Russ Moulton Jr, EOSoft Inc. for his "rad.c" module.
+ */
+/** @} */
+
 #include "common.h"
 
 #include <math.h>
 #include <string.h>
 #include "bio.h"
 
-#include "bu/debug.h"
-#include "bu/log.h"
-#include "bu/malloc.h"
-#include "bu/parallel.h"
 #include "vmath.h"
-#include "bn/tabdata.h"
+#include "bu.h"
+#include "bn.h"
 
 void
 bn_table_free(struct bn_table *tabp)
@@ -75,9 +94,9 @@ bn_ck_table(const struct bn_table *tabp)
 struct bn_table *
 bn_table_make_uniform(size_t num, double first, double last) {
     struct bn_table	*tabp;
-    fastf_t		*fp;
-    fastf_t		delta;
-    size_t		j;
+    fastf_t			*fp;
+    fastf_t			delta;
+    int			j;
 
     if (bu_debug&BU_DEBUG_TABDATA) bu_log("bn_table_make_uniform( num=%zu, %g, %g )\n", num, first, last);
 
@@ -100,7 +119,7 @@ bn_table_make_uniform(size_t num, double first, double last) {
 void
 bn_tabdata_add(struct bn_tabdata *out, const struct bn_tabdata *in1, const struct bn_tabdata *in2)
 {
-    register size_t j;
+    register int		j;
     register fastf_t	*op;
     register const fastf_t	*i1, *i2;
 
@@ -126,7 +145,7 @@ bn_tabdata_add(struct bn_tabdata *out, const struct bn_tabdata *in1, const struc
 void
 bn_tabdata_mul(struct bn_tabdata *out, const struct bn_tabdata *in1, const struct bn_tabdata *in2)
 {
-    register size_t j;
+    register int		j;
     register fastf_t	*op;
     register const fastf_t	*i1, *i2;
 
@@ -152,7 +171,7 @@ bn_tabdata_mul(struct bn_tabdata *out, const struct bn_tabdata *in1, const struc
 void
 bn_tabdata_mul3(struct bn_tabdata *out, const struct bn_tabdata *in1, const struct bn_tabdata *in2, const struct bn_tabdata *in3)
 {
-    register size_t j;
+    register int		j;
     register fastf_t	*op;
     register const fastf_t	*i1, *i2, *i3;
 
@@ -180,7 +199,7 @@ bn_tabdata_mul3(struct bn_tabdata *out, const struct bn_tabdata *in1, const stru
 void
 bn_tabdata_incr_mul3_scale(struct bn_tabdata *out, const struct bn_tabdata *in1, const struct bn_tabdata *in2, const struct bn_tabdata *in3, register double scale)
 {
-    register size_t j;
+    register int		j;
     register fastf_t	*op;
     register const fastf_t	*i1, *i2, *i3;
 
@@ -207,7 +226,7 @@ bn_tabdata_incr_mul3_scale(struct bn_tabdata *out, const struct bn_tabdata *in1,
 void
 bn_tabdata_incr_mul2_scale(struct bn_tabdata *out, const struct bn_tabdata *in1, const struct bn_tabdata *in2, register double scale)
 {
-    register size_t j;
+    register int		j;
     register fastf_t	*op;
     register const fastf_t	*i1, *i2;
 
@@ -232,7 +251,7 @@ bn_tabdata_incr_mul2_scale(struct bn_tabdata *out, const struct bn_tabdata *in1,
 void
 bn_tabdata_scale(struct bn_tabdata *out, const struct bn_tabdata *in1, register double scale)
 {
-    register size_t j;
+    register int		j;
     register fastf_t	*op;
     register const fastf_t	*i1;
 
@@ -256,7 +275,7 @@ bn_tabdata_scale(struct bn_tabdata *out, const struct bn_tabdata *in1, register 
 void
 bn_table_scale(struct bn_table *tabp, register double scale)
 {
-    register size_t j;
+    register int		j;
     register fastf_t	*op;
 
     if (bu_debug&BU_DEBUG_TABDATA) bu_log("bn_table_scale(%p, %g)\n", (void *)tabp, scale);
@@ -272,7 +291,7 @@ bn_table_scale(struct bn_table *tabp, register double scale)
 void
 bn_tabdata_join1(struct bn_tabdata *out, const struct bn_tabdata *in1, register double scale, const struct bn_tabdata *in2)
 {
-    register size_t j;
+    register int		j;
     register fastf_t	*op;
     register const fastf_t	*i1, *i2;
 
@@ -300,7 +319,7 @@ bn_tabdata_join1(struct bn_tabdata *out, const struct bn_tabdata *in1, register 
 void
 bn_tabdata_join2(struct bn_tabdata *out, const struct bn_tabdata *in1, register double scale2, const struct bn_tabdata *in2, register double scale3, const struct bn_tabdata *in3)
 {
-    register size_t j;
+    register int		j;
     register fastf_t	*op;
     register const fastf_t	*i1, *i2, *i3;
 
@@ -331,7 +350,7 @@ bn_tabdata_join2(struct bn_tabdata *out, const struct bn_tabdata *in1, register 
 void
 bn_tabdata_blend2(struct bn_tabdata *out, register double scale1, const struct bn_tabdata *in1, register double scale2, const struct bn_tabdata *in2)
 {
-    register size_t j;
+    register int		j;
     register fastf_t	*op;
     register const fastf_t	*i1, *i2;
 
@@ -359,7 +378,7 @@ bn_tabdata_blend2(struct bn_tabdata *out, register double scale1, const struct b
 void
 bn_tabdata_blend3(struct bn_tabdata *out, register double scale1, const struct bn_tabdata *in1, register double scale2, const struct bn_tabdata *in2, register double scale3, const struct bn_tabdata *in3)
 {
-    register size_t j;
+    register int		j;
     register fastf_t	*op;
     register const fastf_t	*i1, *i2, *i3;
 
@@ -393,7 +412,7 @@ bn_tabdata_area1(const struct bn_tabdata *in)
 {
     register fastf_t		area;
     register const fastf_t	*ip;
-    register size_t j;
+    register int		j;
 
     BN_CK_TABDATA(in);
 
@@ -413,19 +432,17 @@ bn_tabdata_area2(const struct bn_tabdata *in)
     const struct bn_table	*tabp;
     register fastf_t		area;
     fastf_t			width;
-    register size_t j;
+    register int		j;
 
     BN_CK_TABDATA(in);
     tabp = in->table;
     BN_CK_TABLE(tabp);
 
     area = 0;
-    for (j = in->ny-1; j > 0; j--)  {
+    for (j = in->ny-1; j >= 0; j--)  {
 	width = tabp->x[j+1] - tabp->x[j];
 	area += in->y[j] * width;
     }
-    width = tabp->x[1] - tabp->x[0];
-    area += in->y[0] * width;
 
     if (bu_debug&BU_DEBUG_TABDATA) bu_log("bn_tabdata_area2(%p) = %g\n", (void *)in, area);
     return area;
@@ -436,7 +453,7 @@ bn_tabdata_mul_area1(const struct bn_tabdata *in1, const struct bn_tabdata *in2)
 {
     register fastf_t		area;
     register const fastf_t	*i1, *i2;
-    register size_t j;
+    register int		j;
 
     BN_CK_TABDATA(in1);
     BN_CK_TABDATA(in2);
@@ -457,7 +474,7 @@ bn_tabdata_mul_area2(const struct bn_tabdata *in1, const struct bn_tabdata *in2)
     const struct bn_table	*tabp;
     register fastf_t		area;
     fastf_t			width;
-    register size_t j;
+    register int		j;
 
     BN_CK_TABDATA(in1);
     BN_CK_TABDATA(in2);
@@ -465,18 +482,17 @@ bn_tabdata_mul_area2(const struct bn_tabdata *in1, const struct bn_tabdata *in2)
     BN_CK_TABLE(tabp);
 
     area = 0;
-    for (j = in1->ny-1; j > 0; j--)  {
+    for (j = in1->ny-1; j >= 0; j--)  {
 	width = tabp->x[j+1] - tabp->x[j];
 	area += in1->y[j] * in2->y[j] * width;
     }
-    width = tabp->x[1] - tabp->x[0];
-    area += in1->y[0] * in2->y[0] * width;
 
     if (bu_debug&BU_DEBUG_TABDATA) bu_log("bn_tabdata_mul_area2(%p, %p) = %g\n", (void *)in1, (void *)in2, area);
     return area;
 }
 
 /*
+ *			B N _ T A B L E _ F I N D _ X
  *@brief
  *  Return the index in the table's x[] array of the interval which
  *  contains 'xval'.
@@ -493,25 +509,19 @@ bn_tabdata_mul_area2(const struct bn_tabdata *in1, const struct bn_tabdata *in2)
 long
 bn_table_find_x(const struct bn_table *tabp, double xval)
 {
-    register size_t i;
+    register int	i;
 
-    if (bu_debug&BU_DEBUG_TABDATA)
-	bu_log("bn_table_find_x(%p, %g)\n", (void *)tabp, xval);
+    if (bu_debug&BU_DEBUG_TABDATA) bu_log("bn_table_find_x(%p, %g)\n", (void *)tabp, xval);
     BN_CK_TABLE(tabp);
 
-    if (xval > tabp->x[tabp->nx])
-	return -2;
-    if (xval >= tabp->x[tabp->nx-1])
-	return (int)tabp->nx-1;
+    if (xval > tabp->x[tabp->nx])  return -2;
+    if (xval >= tabp->x[tabp->nx-1])  return tabp->nx-1;
 
     /* Search for proper interval in input spectrum */
-    for (i = tabp->nx-2; i > 0; i--)  {
-	if (xval >= tabp->x[i])
-	    return i;
+    for (i = tabp->nx-2; i >=0; i--)  {
+	if (xval >= tabp->x[i])  return i;
     }
-    if (xval >= tabp->x[0])
-	return 0;
-
+    /* if ( xval < tabp->x[0] )  return -1; */
     return -1;
 }
 
@@ -600,10 +610,11 @@ bn_tabdata_resample_max(const struct bn_table *newtable, const struct bn_tabdata
 	     *  Simple case, ends of output span are completely
 	     *  contained within one input span.
 	     *  Interpolate for both ends, take max.
+	     *  XXX this could be more efficiently written inline here.
 	     */
 	    newsamp->y[i] = bn_table_lin_interp(olddata, newtable->x[i]);
 	    tmp = bn_table_lin_interp(olddata, newtable->x[i+1]);
-	    V_MAX(newsamp->y[i], tmp);
+	    if (tmp > newsamp->y[i])  newsamp->y[i] = tmp;
 	} else {
 	    register fastf_t tmp, n;
 	    register int	s;
@@ -615,8 +626,7 @@ bn_tabdata_resample_max(const struct bn_table *newtable, const struct bn_tabdata
 	     */
 	    n = bn_table_lin_interp(olddata, newtable->x[i]);
 	    tmp = bn_table_lin_interp(olddata, newtable->x[i+1]);
-	    V_MAX(n, tmp);
-
+	    if (tmp > n)  n = tmp;
 	    for (s = j+1; s <= k; s++)  {
 		if ((tmp = olddata->y[s]) > n)
 		    n = tmp;
@@ -658,6 +668,7 @@ bn_tabdata_resample_avg(const struct bn_table *newtable, const struct bn_tabdata
 	     *  Simple case, ends of output span are completely
 	     *  contained within one input span.
 	     *  Interpolate for both ends, take average.
+	     *  XXX this could be more efficiently written inline here.
 	     */
 	    newsamp->y[i] = 0.5 * (
 				bn_table_lin_interp(olddata, newtable->x[i]) +
@@ -665,7 +676,7 @@ bn_tabdata_resample_avg(const struct bn_table *newtable, const struct bn_tabdata
 	} else {
 	    /*
 	     *  Complex case: find average value.
-	     *  Interpolate both ends, and consider all
+	     *  Interpolate both end, and consider all
 	     *  intermediate old spans.
 	     *  There are three parts to sum:
 	     *	Partial interval from newx[i] to j+1
@@ -922,14 +933,13 @@ struct bn_tabdata *
 bn_tabdata_binary_read(const char *filename, size_t num, const struct bn_table *tabp) {
     struct bn_tabdata	*data;
     char	*cp;
-    size_t nbytes;
-    size_t len;
-    int got;
+    int	nbytes;
+    int	len;
+    int	got;
     int	fd;
-    size_t i;
+    long i;
 
-    if (bu_debug&BU_DEBUG_TABDATA)
-	bu_log("bn_tabdata_binary_read(%s, num=%zu, %p)\n", filename, num, (void *)tabp);
+    if (bu_debug&BU_DEBUG_TABDATA) bu_log("bn_tabdata_binary_read(%s, num=%zu, %p)\n", filename, num, (void *)tabp);
 
     BN_CK_TABLE(tabp);
 
@@ -951,12 +961,12 @@ bn_tabdata_binary_read(const char *filename, size_t num, const struct bn_table *
     bu_semaphore_acquire(BU_SEM_SYSCALL);
     got = read(fd, (char *)data, len);
     bu_semaphore_release(BU_SEM_SYSCALL);
-    if ((size_t)got != len)  {
+    if (got != len)  {
 	if (got < 0) {
 	    perror(filename);
 	    bu_log("bn_tabdata_binary_read read error on \"%s\"\n", filename);
 	} else {
-	    bu_log("bn_tabdata_binary_read(%s) expected %zu got %d\n", filename, len, got);
+	    bu_log("bn_tabdata_binary_read(%s) expected %d got %d\n", filename, len, got);
 	}
 	bu_free(data, "bn_tabdata[]");
 	bu_semaphore_acquire(BU_SEM_SYSCALL);
@@ -970,14 +980,7 @@ bn_tabdata_binary_read(const char *filename, size_t num, const struct bn_table *
 
     /* Connect data[i].table pointer to tabp */
     cp = (char *)data;
-    for (i = num-1; i > 0; i--, cp += nbytes)  {
-	register struct bn_tabdata *sp;
-	sp = (struct bn_tabdata *)cp;
-	BN_CK_TABDATA(sp);
-	sp->table = tabp;
-    }
-    {
-	/* i == 0 */
+    for (i = num-1; i >= 0; i--, cp += nbytes)  {
 	register struct bn_tabdata *sp;
 	sp = (struct bn_tabdata *)cp;
 	BN_CK_TABDATA(sp);
@@ -1049,16 +1052,15 @@ bn_tabdata_dup(const struct bn_tabdata *in) {
 struct bn_tabdata *
 bn_tabdata_get_constval(double val, const struct bn_table *tabp) {
     struct bn_tabdata	*data;
-    size_t		todo;
+    int			todo;
     register fastf_t	*op;
 
     BN_CK_TABLE(tabp);
     BN_GET_TABDATA(data, tabp);
 
     op = data->y;
-    for (todo = data->ny-1; todo > 0; todo--)
+    for (todo = data->ny-1; todo >= 0; todo--)
 	*op++ = val;
-    *op++ = val; /* todo == 0 */
 
     if (bu_debug&BU_DEBUG_TABDATA) bu_log("bn_tabdata_get_constval(val=%g, %p)=%p\n", val, (void *)tabp, (void *)data);
 
@@ -1068,7 +1070,7 @@ bn_tabdata_get_constval(double val, const struct bn_table *tabp) {
 void
 bn_tabdata_constval(struct bn_tabdata *data, double val)
 {
-    size_t		todo;
+    int			todo;
     register fastf_t	*op;
 
     if (bu_debug&BU_DEBUG_TABDATA) bu_log("bn_tabdata_constval(%p, val=%g)\n", (void *)data, val);
@@ -1076,9 +1078,8 @@ bn_tabdata_constval(struct bn_tabdata *data, double val)
     BN_CK_TABDATA(data);
 
     op = data->y;
-    for (todo = data->ny-1; todo > 0; todo--)
+    for (todo = data->ny-1; todo >= 0; todo--)
 	*op++ = val;
-    *op++ = val; /* todo == 0 */
 }
 
 void
@@ -1103,8 +1104,8 @@ bn_tabdata_to_tcl(struct bu_vls *vp, const struct bn_tabdata *data)
     for (i=0; i < data->ny; i++)  {
 	register fastf_t val = data->y[i];
 	bu_vls_printf(vp, "%g ", val);
-	V_MIN(minval, val);
-	V_MAX(maxval, val);
+	if (val < minval)  minval = val;
+	if (val > maxval)  maxval = val;
     }
     bu_vls_printf(vp, "} nx %zu ymin %g ymax %g",
 		  tabp->nx, minval, maxval);
@@ -1162,7 +1163,7 @@ bn_tabdata_freq_shift(struct bn_tabdata *out, const struct bn_tabdata *in, doubl
     }
 }
 
-size_t
+int
 bn_table_interval_num_samples(const struct bn_table *tabp, double low, double hi)
 {
     register size_t i;
@@ -1178,14 +1179,13 @@ bn_table_interval_num_samples(const struct bn_table *tabp, double low, double hi
     return count;
 }
 
-size_t
-bn_table_delete_sample_pts(struct bn_table *tabp, size_t i, size_t j)
+int
+bn_table_delete_sample_pts(struct bn_table *tabp, unsigned int i, unsigned int j)
 {
     size_t tokill;
-    size_t k;
+    unsigned int k;
 
-    if (bu_debug&BU_DEBUG_TABDATA)
-	bu_log("bn_table_delete_samples(%p, %d, %d)\n", (void *)tabp, i, j);
+    if (bu_debug&BU_DEBUG_TABDATA) bu_log("bn_table_delete_samples(%p, %d, %d)\n", (void *)tabp, i, j);
 
     BN_CK_TABLE(tabp);
 
@@ -1193,11 +1193,8 @@ bn_table_delete_sample_pts(struct bn_table *tabp, size_t i, size_t j)
 	bu_bomb("bn_table_delete_sample_pts() index out of range\n");
 
     tokill = j - i + 1;
-
-    if (tokill < 1)
-	bu_bomb("bn_table_delete_sample_pts(): nothing to delete\n");
-    if (tokill >= tabp->nx)
-	bu_bomb("bn_table_delete_sample_pts(): you can't kill 'em all!\n");
+    if (tokill < 1)  bu_bomb("bn_table_delete_sample_pts(): nothing to delete\n");
+    if (tokill >= tabp->nx) bu_bomb("bn_table_delete_sample_pts(): you can't kill 'em all!\n");
 
     tabp->nx -= tokill;
 
@@ -1242,8 +1239,7 @@ bn_table_merge2(const struct bn_table *a, const struct bn_table *b) {
 	    newtab->x[k++] = b->x[j++];
 	}
     }
-    if (k > newtab->nx)
-	bu_bomb("bn_table_merge2() assertion failed, k>nx?\n");
+    if (k > newtab->nx)  bu_bomb("bn_table_merge2() assertion failed, k>nx?\n");
     newtab->nx = k-1;
 
     if (bu_debug&BU_DEBUG_TABDATA) bu_log("bn_table_merge2(%p, %p) = %p\n", (void *)a, (void *)b, (void *)newtab);
@@ -1288,7 +1284,7 @@ bn_tabdata_mk_linear_filter(const struct bn_table *spectrum, double lower_wavele
 	       upper_wavelen, spectrum->x[0]);
 	return NULL;
     } else if (last <= -2)  {
-	last = (int)spectrum->nx-1;
+	last = spectrum->nx-1;
 	bu_log("bn_tabdata_mk_linear_filter() warning, upper_wavelen %g > highest sampled wavelen %g\n",
 	       upper_wavelen, spectrum->x[spectrum->nx]);
     }
@@ -1313,16 +1309,14 @@ bn_tabdata_mk_linear_filter(const struct bn_table *spectrum, double lower_wavele
     filt_range = spectrum->x[first+1] - lower_wavelen;
     cell_range = spectrum->x[first+1] - spectrum->x[first];
     frac = filt_range / cell_range;
-    V_MIN(frac, 1);
-
+    if (frac > 1)  frac = 1;
     BU_ASSERT((frac >= 0.0) && (frac <= 1.0));
     filt->y[first] = frac;
 
     filt_range = upper_wavelen - spectrum->x[last];
     cell_range = spectrum->x[last+1] - spectrum->x[last];
     frac = filt_range / cell_range;
-    V_MIN(frac, 1);
-
+    if (frac > 1)  frac = 1;
     BU_ASSERT((frac >= 0.0) && (frac <= 1.0));
     filt->y[last] = frac;
 
@@ -1332,6 +1326,7 @@ bn_tabdata_mk_linear_filter(const struct bn_table *spectrum, double lower_wavele
 
     return filt;
 }
+/** @} */
 /*
  * Local Variables:
  * mode: C
